cleanup(settings): Cleaned up character & campaign settings
