+++ src/api-calls/character-campaign-settings/listenToSettings.ts
@@ -1,11 +1,11 @@
 import { onSnapshot } from "firebase/firestore";
-import { SettingsDoc } from "types/Settings.type";
+import { SettingsDocument } from "api-calls/character-campaign-settings/_character-campaign-settings.type";
 import { getCampaignSettingsDoc, getCharacterSettingsDoc } from "./_getRef";
 
 export function listenToSettings(
   campaignId: string | undefined,
   characterId: string | undefined,
-  onSettings: (settings: SettingsDoc) => void,
+  onSettings: (settings: SettingsDocument) => void,
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   onError: (error: any) => void
 ) {
