+++ src/api-calls/character-campaign-settings/showOrHideCustomOracle.ts
